+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - links to the API from the exmaples ([#185])
+- benchmark test via codspeed.io ([#188])
 
 [#185]: https://github.com/dschick/udkm1Dsim/pull/185
+[#188]: https://github.com/dschick/udkm1Dsim/pull/188
 
 ### Changed